requirement.txt
ultralytics
lap
flask
pywebview
